requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
asyncer==0.0.7
bidict==0.23.1
certifi==2025.8.3
chainlit==2.3.0
charset-normalizer==3.4.2
chevron==0.14.0
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.14
filetype==1.2.0
googleapis-common-protos==1.70.0
griffe==1.11.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
Lazify==0.4.0
literalai==0.1.103
logfire-api==4.2.0
marshmallow==3.26.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
openai==1.99.3
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
protobuf==6.31.1
pydantic==2.11.7
pydantic-ai-slim==0.6.2
pydantic-graph==0.6.2
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
requests==2.32.4
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.3
syncer==2.0.3
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uptrace==1.35.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.23.0
